Add PegasusToken and use it for existing cards. Implement card: Pegasus Refuge
